Added singleton class to query database and removed couchbase direct references from MainActivity
