--- earthviewwallpaper/src/main/java/com/github/p1va/earthviewwallpaper/data/model/GoogleEarthViewImage.java
@@ -1,29 +1,0 @@
-package com.github.p1va.earthviewwallpaper.data.model;
-
-/**
- * Created by Stefano Piva on 24/09/2017.
- */
-
-public class GoogleEarthViewImage {
-    public String id;
-    public String api;
-    public String attribution;
-    public String country;
-    public String downloadUrl;
-    public String earthLink;
-    public String earthTitle;
-    public String lat;
-    public String lng;
-    public String mapsLink;
-    public String mapsTitle;
-    public String nextApi;
-    public String nextUrl;
-    public String photoUrl;
-    public String prevApi;
-    public String prevUrl;
-    public String region;
-    public String slug;
-    public String thumbUrl;
-    public String title;
-    public String url;
-}
