+++ earthviewwallpaper/src/main/res/layout/item_image_preview.xml
@@ -21,13 +21,13 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
-        android:background="#A6000000"
+        android:background="#70000000"
         android:textColor="#FFFFFF"
         android:gravity="start"
-        android:paddingLeft="4dp"
-        android:paddingStart="4dp"
-        android:paddingRight="4dp"
-        android:paddingEnd="4dp"
+        android:paddingLeft="8dp"
+        android:paddingStart="8dp"
+        android:paddingRight="8dp"
+        android:paddingEnd="8dp"
         android:paddingTop="8dp"
         android:paddingBottom="8dp"
         android:textSize="16sp"/>
